requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.1
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.3
duplicity==0.8.21
exceptiongroup==1.1.3
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
instaloader==4.10.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
openai==1.3.3
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.2.0
postgres==4.0
prompt-toolkit==3.0.30
protobuf==3.12.4
psutil==5.9.6
psycopg2-binary==2.9.3
psycopg2-pool==1.1
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.1
pydantic_core==2.14.3
pygame==2.5.2
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.6
python-dotenv==0.19.2
python-telegram-bot==20.6
pytz==2022.2.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.3.4
reportlab==3.6.8
requests==2.28.1
rich==13.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.2
systemd-python==234
texttable==1.6.4
tqdm==4.66.1
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.11
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
